requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Statistical Models
statsmodels>=0.14.0
pmdarima>=2.0.0

# XAI Libraries
shap>=0.42.0
captum>=0.6.0

# Experiment Tracking
mlflow>=2.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Metrics and Statistics
scipy>=1.11.0

# Data Fetching (for gold prices)
yfinance>=0.2.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Gradient Boosting
xgboost>=1.7.0
lightgbm>=4.0.0

# Reproducibility
joblib>=1.3.0

# Web Application
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Runtime Configuration
python-dotenv>=1.0.0
